Add Datasource\Locator\AbstractLocator
Classes implementing Datasource\Locator\LocatorInterface can extend
this class to avoid code duplication.
